# swagger.yaml
openapi: 3.0.3
info:
  title: Energy Dashboard API
  description: |
    Comprehensive API for Energy Dashboard Management System
    
    This API provides endpoints for:
    - Authentication and user management
    - Chart builder functionality
    - Dashboard layout management
    - Data aggregation and analytics
    - Real-time data streaming
    
    ## Authentication
    All endpoints (except login/register) require JWT authentication.
    Include the token in the Authorization header: `Bearer <token>`
    
    ## Response Format
    All responses follow a consistent format:
    ```json
    {
      "success": true,
      "data": {...},
      "message": "Success message",
      "timestamp": "2025-06-18T18:00:00.000Z"
    }
    ```
    
  version: 1.0.0
  contact:
    name: Energy Dashboard API Support
    email: support@energydashboard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.energydashboard.com/api
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          type: object
          description: Response data
        message:
          type: string
          description: Response message
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        error:
          type: object
          description: Error details (only present when success is false)
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    Dashboard:
      type: object
      properties:
        id:
          type: string
          description: Dashboard ID
        name:
          type: string
          description: Dashboard name
        description:
          type: string
          description: Dashboard description
        level:
          type: string
          enum: [SECTOR, STATE, PLANT, EQUIPMENT]
          description: Dashboard level
        dashboardType:
          type: string
          enum: [LIVE, ANALYTICAL]
          description: Dashboard type
        layout:
          type: object
          properties:
            columns:
              type: integer
              minimum: 1
              maximum: 24
            rows:
              type: integer
              minimum: 1
            gap:
              type: integer
              minimum: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/DashboardWidget'
        tags:
          type: array
          items:
            type: string
        isTemplate:
          type: boolean
        isPublic:
          type: boolean
        permissions:
          type: object
          properties:
            view:
              type: array
              items:
                type: string
            edit:
              type: array
              items:
                type: string
        usageStats:
          type: object
          properties:
            viewCount:
              type: integer
            avgSessionDuration:
              type: number
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DashboardWidget:
      type: object
      properties:
        id:
          type: string
          description: Widget ID
        type:
          type: string
          enum: [chart, metric, table, text, image]
          description: Widget type
        title:
          type: string
          description: Widget title
        position:
          type: object
          properties:
            x:
              type: integer
              minimum: 0
            y:
              type: integer
              minimum: 0
        size:
          type: object
          properties:
            width:
              type: integer
              minimum: 1
            height:
              type: integer
              minimum: 1
        config:
          type: object
          description: Widget-specific configuration
        chartId:
          type: string
          description: Reference to chart (for chart widgets)
        options:
          type: object
          properties:
            showTitle:
              type: boolean
            refreshInterval:
              type: integer
            autoRefresh:
              type: boolean
            customStyles:
              type: object
            resizable:
              type: boolean
            draggable:
              type: boolean
            removable:
              type: boolean
            collapsible:
              type: boolean
            collapsed:
              type: boolean
  parameters:
...
    TimeRangeParam:
      name: timeRange
      in: query
      description: Time range for data
      schema:
        type: string
        enum: [1h, 6h, 24h, 7d, 30d]
        default: 24h
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
            timestamp: "2025-06-18T18:00:00.000Z"
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: "FORBIDDEN"
              message: "Insufficient permissions"
            timestamp: "2025-06-18T18:00:00.000Z"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
            timestamp: "2025-06-18T18:00:00.000Z"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid input data"
              details:
                field: "email"
                message: "Invalid email format"
            timestamp: "2025-06-18T18:00:00.000Z"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: "INTERNAL_ERROR"
              message: "Internal server error"
            timestamp: "2025-06-18T18:00:00.000Z"
paths:
...
  /dashboard-layout/dashboards/{dashboardId}:
    get:
      tags:
        - Dashboard Layout
      summary: Get dashboard by ID
      description: Retrieve a specific dashboard with widget data
      parameters:
        - name: dashboardId
          in: path
          required: true
          description: Dashboard ID
          schema:
            type: string
        - $ref: '#/components/parameters/TimeRangeParam'
        - name: includeData
          in: query
          description: Include widget data in response
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Dashboard retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          dashboard:
                            $ref: '#/components/schemas/Dashboard'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      tags:
        - Dashboard Layout
      summary: Update dashboard
      description: Update an existing dashboard
      parameters:
        - name: dashboardId
          in: path
          required: true
          description: Dashboard ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                level:
                  type: string
                  enum: [SECTOR, STATE, PLANT, EQUIPMENT]
                dashboardType:
                  type: string
                  enum: [LIVE, ANALYTICAL]
                layout:
                  type: object
                  properties:
                    columns:
                      type: integer
                      minimum: 1
                      maximum: 24
                    rows:
                      type: integer
                      minimum: 1
                    gap:
                      type: integer
                      minimum: 0
                tags:
                  type: array
                  items:
                    type: string
                isTemplate:
                  type: boolean
                isPublic:
                  type: boolean
      responses:
        '200':
          description: Dashboard updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Dashboard'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      tags:
        - Dashboard Layout
      summary: Delete dashboard
      description: Delete a dashboard
      parameters:
        - name: dashboardId
          in: path
          required: true
          description: Dashboard ID
          schema:
            type: string
      responses:
        '200':
          description: Dashboard deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "Dashboard deleted successfully"
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'
...